tests/wpt/web-platform-tests/css/css-text/parsing/letter-spacing-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Text Module Test: parsing letter-spacing with valid values</title>
<link rel="help" href="https://www.w3.org/TR/css-text-3/#propdef-letter-spacing">
<meta name="assert" content="letter-spacing supports the full grammar 'normal | <length>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("letter-spacing", "normal");

test_valid_value("letter-spacing", "0", "0px");
test_valid_value("letter-spacing", "-10px");
test_valid_value("letter-spacing", "calc(2em + 3ex)");
</script>
</body>
</html>
